country code notifiable notification grouping fix

+++ main.py
@@ -1,25 +1,39 @@
 #!/usr/bin/env python3
 import datetime
+import logging
 import smtplib
 import sys
+from collections import defaultdict
 from itertools import zip_longest
 
 import requests
+from envelope import envelope
 from flask import Flask, render_template, request, jsonify, redirect, url_for, send_from_directory, flash, session
 from flask_debugtoolbar import DebugToolbarExtension
 from flask_oidc import OpenIDConnect
 from flask_paginate import Pagination
 from flask_session import Session  # new style
 from flask_wtf import FlaskForm
-from sqlalchemy import or_
+from sqlalchemy import or_, and_
 from sqlalchemy.exc import DataError
-from wtforms import StringField, SelectField, SelectMultipleField
+from wtforms import StringField, SelectMultipleField
 
-from mail_controller import build_mail
-from models import PrefixAsn, Conflict, Notification, User, db
+from models import PrefixAsn, Conflict, Notification, User, db, MailHistory
+
+# setup logging
+console_handler = logging.StreamHandler()
+console_handler.setFormatter(logging.Formatter('%(message)s'))
+console_handler.setLevel(logging.INFO)
+file_handler = logging.FileHandler("rpki_sentry.log")
+file_handler.setFormatter(logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s'))
+file_handler.setLevel(logging.WARNING)
+logging.basicConfig(level=logging.INFO, handlers=[console_handler, file_handler])
+logger = logging.getLogger(__name__)
+
 
 app = Flask(__name__)
 
+HOSTNAME = "https://rpki-sentry.csirt.cz/"
 # App config.
 # DEBUG = True
 SESSION_TYPE = 'sqlalchemy'
@@ -33,7 +47,7 @@ SQLALCHEMY_ECHO = True
 OIDC_CLIENT_SECRETS = '/var/www/html/client_secrets.json'
 OIDC_ID_TOKEN_COOKIE_SECURE = False
 OIDC_REQUIRE_VERIFIED_EMAIL = False
-OIDC_OPENID_REALM = 'https://rpki-sentry.csirt.cz:5002/oidc_callback'
+OIDC_OPENID_REALM = 'https://rpki-sentry.csirt.cz:5002/oidc_callback' # XX why is here port 5002?
 
 DEBUG_TB_INTERCEPT_REDIRECTS = False
 
@@ -106,7 +120,6 @@ def charts():
     h_lab, h_1, h_2 = zip(*db.engine.execute(
         'select ts,sum(case when status = 16 then 1 else 0 end) asnconflict, sum(case when status = 17 then 1 else 0 end) pfxlenconflict from seen_tables inner join conflict on (ts > start and (ts < "end" or "end" is null)) group by ts order by ts asc'))
     h_lab = (str(x.date()) for x in h_lab)
-    # import ipdb; ipdb.set_trace() labels=labels, axe_x=axe_x, axe_y=axe_y
     return render_template('charts.html', h_lab=h_lab, h_1=h_1, h_2=h_2, o_lab=o_lab, o_1=o_1)
 
 
@@ -116,18 +129,17 @@ def notifications():
     if request.method == 'POST':
         # user.email = request.form.get("email")
         db.session.add(user)
-        for n_id, prefix, asn in zip_longest(request.form.getlist("id[]"),
-                                             request.form.getlist("prefix[]"),
-                                             request.form.getlist("asn[]")):
-            print(n_id, prefix, asn)
-            # import ipdb; ipdb.set_trace()
+        for n_id, prefix, asn, ccs in zip_longest(request.form.getlist("id[]"),
+                                                  request.form.getlist("prefix[]"),
+                                                  request.form.getlist("asn[]"),
+                                                  request.form.getlist("ccs[]")):
             if n_id:
                 notification = db.session.query(Notification).filter(Notification.id == n_id,
                                                                      Notification.user_id == user.id).first()
                 if not notification:
                     print("not found", n_id)
                     return jsonify({"success": False, "message": f"Couldn't find notification {n_id}"})
-                if not prefix and not asn:  # remove or update
+                if not prefix and not asn and not ccs:  # remove or update
                     db.session.delete(notification)
                     continue
                 print("found", notification.id)
@@ -137,6 +149,7 @@ def notifications():
                 notification.user_id = user.id
             notification.prefix = prefix or None
             notification.asn = asn or None
+            notification.ccs = ccs or None
             db.session.add(notification)
         try:
             db.session.commit()
@@ -158,11 +171,14 @@ def notifications():
 def add():
     asn = request.args.get('asn')
     prefix = request.args.get('cidr')
+    ccs = request.args.getlist('ccs')
     n = Notification()
     if asn and asn != 'None':
         n.asn = asn
     if prefix and prefix != 'None':
         n.prefix = prefix
+    if ccs and ccs != 'None':
+        n.ccs = ",".join(ccs)
     n.user_id = User.himself(False)
     db.session.add(n)
     try:
@@ -177,7 +193,7 @@ def add():
 @app.route("/search/<int:page>", methods=['GET'])
 def search(page=1):
     if "bgpsec.labs.nic.cz" in request.headers['Host']:  # service should not be available through bgpsec anymore
-        return redirect("https://rpki-sentry.csirt.cz/")
+        return redirect(HOSTNAME)
 
     asn = request.args.get('asn')
     cidr = request.args.get('cidr')
@@ -209,10 +225,18 @@ def search(page=1):
 
     date_from = db.engine.execute("select min(ts) from seen_tables").fetchone()[0]
     date_to = db.engine.execute("select max(ts) from seen_tables").fetchone()[0]
-    form.cc_not.choices = form.cc.choices = [(x[0],x[0].upper()) for x in db.engine.execute("select cc from prefix_asn where cc is not null group by cc")]
+    form.cc_not.choices = form.cc.choices = [(x[0], x[0].upper()) for x in
+                                             db.engine.execute("select cc from prefix_asn where cc is not null group by cc")]
 
     return render_template('search.html', form=form, data=data, date_from=date_from, date_to=date_to,
                            pagination=pagination)
+
+
+@app.template_filter('split')
+def split(value, char=','):
+    if not value:
+        return []
+    return value.split(char)
 
 
 @app.template_filter('status')
@@ -232,90 +256,114 @@ def status_2_word(s):
 
 if __name__ == "__main__":
     if "process" in sys.argv:
-        # XX
-        #
-        #  odeslat maily o změnách (zmíněné prefix_asn_id) (datetime start | datetime_end) > poslední běh
-        #    projdi vsechny notifikace a pro kazdou vyhledej v novych radcich dvojice
-        #       kde conflicts/asn podprefix   odpovida notifikace/asn+ nadprefix
-        #   a zavolej send_new_conflict()
-        #  přegenerovat zdroj pro dat grafů (pro chartjs)
-        #  odstranit starší než půlroku z DB
-        # import ipdb; ipdb.set_trace()
+        #  * load country codes for new prefixes
+        #  * loop all notification and search for conflicts/asn subprefix relates to notification/asn superprefix
+        #  * send notification e-mails
+        #  * todo: XXregenerate chart sources (for chartjs)
+        #  * todo: XXremove older than six months from DB?
+        no_countries = "mails" in sys.argv  # for debug
 
         app.app_context().push()
         db.engine.echo = False  # turn off default stdout logging
-        # with create_engine(SQLALCHEMY_DATABASE_URI).connect() as con:
         last_check = db.session.execute("select progress_time from state where id = '1'").fetchone()[0]
-
         if not last_check:
             # fallback is the oldest date we know
             last_check = db.engine.execute("select min(ts) from seen_tables").fetchone()[0]
         now = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')
 
         # add missing country codes
-        CONVEY_URL = "http://127.0.0.1:26683/?q={}"
+        prefix_asn_newly_fetched = []
+        if not no_countries:
+            CONVEY_URL = "http://127.0.0.1:26683/?q={}&field=country"
 
-        # not having country code set, start with new ID (don't get stuck with an erroneous whois-records from last time)
-        q = db.session.query(PrefixAsn)\
-            .filter(PrefixAsn.cc == None)\
-            .order_by(PrefixAsn.id.desc())
-        for i, prefix in enumerate(q.all()):
-            url = CONVEY_URL.format(prefix.prefix)
-            r = requests.get(url=url, timeout=310)
-            cc = r.json()["country"]
-            if cc:
-                if len(cc) == 2:
-                    prefix.cc = cc
-                else:
-                    print("Wrong cc length:", prefix.prefix, cc)
-            print(url, prefix.cc)
-            if i % 10 == 9:
-                print("Comitting")
-                db.session.commit()
-        db.session.commit()
+            # not having country code set, start with new ID (don't get stuck with an erroneous whois-records from last time)
+            # XX what if an IP change its country after few months?
+            q = db.session.query(PrefixAsn) \
+                .filter(PrefixAsn.cc == None) \
+                .order_by(PrefixAsn.id.desc())
+            for i, prefix_asn_ in enumerate(q.all()):
+                url = CONVEY_URL.format(prefix_asn_.prefix)
+                r = requests.get(url=url, timeout=310)
+                cc = r.json()["country"]
+                if cc:
+                    if len(cc) == 2:
+                        prefix_asn_.cc = cc
+                        prefix_asn_newly_fetched.append(prefix_asn_.id)
+                    else:
+                        print("Wrong cc length:", prefix_asn_.prefix, cc)
+                print(url, prefix_asn_.cc)
+                if i % 10 == 9:
+                    print("Comitting")
+                    db.session.commit()
+            db.session.commit()
 
-        # loop all notification and the new lines
+        # build the query that checks if there exists a new conflicts we should be told about
+        conflicts = defaultdict(list)  # [email] => [conflicts...]
+
+        # condition
+        conditions = [Conflict.end > last_check, Conflict.start > last_check]  # must be new
+        if prefix_asn_newly_fetched:  # some old conflicts might be updated
+            # This prefixes could be processed in the past and skipped because country code has not been known
+            # Since it is known now, notify about it eve if it was processed before than last check.
+            conditions.append(and_(PrefixAsn.id.in_(prefix_asn_newly_fetched), Notification.ccs.isnot(None)))
+        query = (db.session.query(PrefixAsn, Conflict, Notification, User)
+                 .join(Conflict, Conflict.prefix_asn_id == PrefixAsn.id)
+                 .join(Notification, and_(
+            # map conflicts to notification if all conditions are met
+            or_(Notification.prefix.is_(None), PrefixAsn.prefix.op("<<=")(Notification.prefix)),
+            or_(Notification.asn.is_(None), PrefixAsn.asn == Notification.asn),
+            or_(Notification.ccs.is_(None), PrefixAsn.cc.contains(Notification.ccs))
+        ))
+                 .join(User, User.id == Notification.user_id)
+                 .filter(User.email.isnot(None))  # there must be an e-mail to send it to
+                 .filter(or_(*conditions))  # conflict must be new or country-updated
+                 )
+        for row in query.all():
+            conflicts[row.User.id].append((row.Notification, row.PrefixAsn, row.Conflict, row.User))
+
+        # send e-mails
         smtp_server = "localhost"
-        with smtplib.SMTP(smtp_server) as smtp:
-            for grouped in db.session.query(Notification, User) \
-                    .join(User, User.id == Notification.user_id) \
-                    .filter(User.email.isnot(None)) \
-                    .group_by(Notification.id, User.id, Notification.user_id).all():
-                conflicts = []
-                user = grouped.User
-                for notification in db.session.query(Notification).filter(Notification.user_id == user.id).all():
-                    query = db.session.query(PrefixAsn, Conflict) \
-                        .join(Conflict, Conflict.prefix_asn_id == PrefixAsn.id) \
-                        .filter(or_(Conflict.end > last_check, Conflict.start > last_check))
-                    if notification.asn:
-                        query = query.filter(PrefixAsn.asn == notification.asn)
-                    if notification.prefix:
-                        query = query.filter(PrefixAsn.prefix.op("<<=")(notification.prefix))
-                    for conflict in query.all():
-                        conflicts.append((notification.asn, notification.prefix,
-                                          conflict.PrefixAsn.prefix, conflict.Conflict.start, conflict.Conflict.end))
-
-                if conflicts:  # notify user
+        if conflicts:  # notify user
+            with smtplib.SMTP(smtp_server) as smtp:
+                for items in conflicts.values():
                     # notify(user_id, conflicts)
                     res = []
-                    for c in conflicts:
-                        res.append(f"Watched ASN: {c[0]}<br>" \
-                                       f"Watched prefix: {c[1]}<br>" \
-                                       f"Matched prefix: {c[2]}<br>" \
-                                       f"Conflict start: {c[3]}<br>" \
-                                       f"Conflict end: {c[4]}<br>")
+                    for notification, prefix_asn, conflict, user in items:
+                        l = []
+                        if notification.asn:
+                            l.append(f"Watched ASN: {notification.asn}<br>")
+                        if notification.prefix:
+                            l.append(f"Watched prefix: {notification.prefix}<br>")
+                        if notification.ccs:
+                            l.append(f"Watched countries: {notification.ccs}<br>")
+                        l.append(f"Matched prefix: {prefix_asn.prefix}<br>"
+                                 f"Conflict start: {conflict.start}<br>"
+                                 f"Conflict end: {conflict.end}<br>")
+                        res.append("".join(l))
 
                     body = "<br>----<br>".join(res)
-                    build_mail(user.email, "RPKI Sentry notification", body, smtp, False)
-                    # Celkem sledujete tyto radky:
-                    # + get_notification_text
-                    # unsubscribe nebo změna proběhne tam
-                    # ... _send_email()
-                    # mail history uloz radek s current_btimestamp
-                    # XX mail history?
+                    body += f"<br><br>Go to <a href={HOSTNAME}/notifications>notification</a> page to edit notification."
 
-        # XX db.session.execute("update state SET progress_time = '" + now + "' where id = '1' ")
-        # db.session.commit()
+                    success = (envelope()
+                               .sender("rpki-sentry@csirt.cz")
+                               # XXX .to(user.email)
+                               .to("edvard.rejthar+test@nic.cz")
+                               .subject("RPKI Sentry notification")
+                               .smtp("smtp.ini")
+                               .message(body)
+                               .signature()
+                               .send(False))
+                    # XX add unsubscribe link
+                    print(success)
+                    if success:
+                        # make a note to the DB that we've sent this mail
+                        db.session.merge(MailHistory(user_id=user.id, timestamp=now))
+                    else:
+                        logger.error(f"Mail not sent {user.email}, body: {body}")
+
+        # XXX after testing
+        # db.session.execute("update state SET progress_time = '" + now + "' where id = '1' ")
+        db.session.commit()
     else:
         # app.run(ssl_context=('/etc/letsencrypt/live/rpki-sentry.csirt.cz/fullchain.pem', '/etc/letsencrypt/live/rpki-sentry.csirt.cz/privkey.pem'))
         # app.run(ssl_context='adhoc')
